refactor(template): literal lib/config.ts as the single source of truth
Replace the env-driven, two-layer config with a literal typed object in
lib/config.ts so the template's behavior is legible in one place. Forks
edit code; .env holds only credentials.

- Move shop.config.ts to lib/config.ts as a literal config; drop env-var
  feature flags (014999ee, 20ab23da).
- Move default nav/footer menus out of config into lib/navigation.ts and
  inline nav/footer/social defaults into their components (603b1965,
  1a70b582).
- Slim .env.example to required-first, alphabetical, sample values and
  drop decorative headings (7f671145..d026e3e5).
- Point the enable-analytics skill at lib/config.ts (0f6fda07).
- Remove the underbaked CMS draft mode (c3e97981).
- Prune dead turbo globalEnv; use server VERCEL_PROJECT_PRODUCTION_URL
  (3b67cb4b).
- Make NEXT_PUBLIC_BASE_URL a bare domain to match Vercel's URL system
  vars (4c013b9d).
