+++ apps/docs/content/docs/anatomy/pages/pdp.mdx
@@ -27,7 +27,7 @@ The Product Detail Page (PDP) is a single product page — title, price, media, 
 - **Related products** — algorithmic "you might also like" set below the product details.
 - **Structured data** — Product and Breadcrumb JSON-LD for search and social.
 
-The quantity picker, Buy with Shop, bundles, complementary products, and related products are on by default and toggled under `pdp` in `shop.config.ts`.
+The quantity picker, Buy with Shop, bundles, complementary products, and related products are on by default and toggled under `pdp` in `lib/config.ts`.
 
 ## Common customizations
