+++ apps/docs/content/docs/anatomy/pages/plp.mdx
@@ -24,4 +24,4 @@ The Product Listing Page (PLP) covers two routes — `/collections/[handle]` for
 - **Predictive search** — suggestion chips and product results as the user types, from the nav search modal.
 - **Structured data** — Collection JSON-LD on collection pages.
 
-None of the PLP's features are toggled in `shop.config.ts` today — they ship on for every storefront.
+None of the PLP's features are toggled in `lib/config.ts` today — they ship on for every storefront.
