+++ apps/docs/content/docs/reference/routes.mdx
@@ -25,7 +25,6 @@ type: reference
 | Route                        | Description                                    |
 | ---------------------------- | ---------------------------------------------- |
 | `POST /api/webhooks/shopify` | Shopify webhook handler for cache invalidation |
-| `GET /api/draft`             | Draft mode toggle for content preview          |
 | `POST /api/chat`             | Guarded AI SDK shopping-assistant stream       |
 
 ## Variant URLs
