+++ apps/docs/content/docs/skills/enable-analytics.mdx
@@ -16,7 +16,7 @@ type: guide
 
 # Enable Analytics
 
-The current storefront includes support for Vercel Web Analytics and Vercel Speed Insights, with each integration disabled by default in `shop.config.ts`. This skill enables or adds those integrations and can also add Google Tag Manager using the recommended integration.
+The current storefront includes support for Vercel Web Analytics and Vercel Speed Insights, with each integration disabled by default in `lib/config.ts`. This skill enables or adds those integrations and can also add Google Tag Manager using the recommended integration.
 
 ## Before you start
 
@@ -39,7 +39,7 @@ Wait for the user to answer both questions before proceeding.
 
 ## Part A: Vercel Analytics and Speed Insights
 
-If the storefront has `analytics` configuration in `shop.config.ts`, enable only the selected integrations. If the user selected neither, keep both integration gates disabled and skip the remaining steps in this section.
+If the storefront has `analytics` configuration in `lib/config.ts`, enable only the selected integrations. If the user selected neither, keep both integration gates disabled and skip the remaining steps in this section.
 
 ```ts
 analytics: {
@@ -90,7 +90,7 @@ Set the actual value in `.env.local` or in the Vercel dashboard under Environmen
 
 ### B3. Add GTM to `components/analytics.tsx`
 
-Import `GoogleTagManager` from `@next/third-parties/google`. Read `NEXT_PUBLIC_GTM_ID` in the analytics component and render `<GoogleTagManager gtmId={gtmId} />` only when the value exists. If the storefront extends `shop.config.ts` with a GTM integration gate, apply that gate inside the same component.
+Import `GoogleTagManager` from `@next/third-parties/google`. Read `NEXT_PUBLIC_GTM_ID` in the analytics component and render `<GoogleTagManager gtmId={gtmId} />` only when the value exists. If the storefront extends `lib/config.ts` with a GTM integration gate, apply that gate inside the same component.
 
 ---
 
@@ -105,7 +105,7 @@ import { GoogleTagManager } from "@next/third-parties/google";
 import { Analytics } from "@vercel/analytics/next";
 import { SpeedInsights } from "@vercel/speed-insights/next";
 
-import { shopConfig } from "@/shop.config";
+import { shopConfig } from "@/lib/config";
 
 export function AnalyticsComponents() {
   const gtmId = process.env.NEXT_PUBLIC_GTM_ID;
