+++ apps/docs/content/docs/skills/enable-shopify-menus.mdx
@@ -16,7 +16,7 @@ type: guide
 
 # Enable Shopify Menus
 
-By default, the storefront's nav and footer render hardcoded items from `shopConfig.navigation` in `shop.config.ts`. The components themselves (`QuickLinks`, `MobileMenu`, `Footer`'s `FooterMenu`) already consume `MenuItem[]` in Shopify shape and render up to three levels of nesting. This skill swaps the data source from `shopConfig.navigation.nav` / `shopConfig.navigation.footer` to a live Shopify menu fetched by handle, with a fallback to the hardcoded items if the menu is missing or empty.
+By default, the storefront's nav and footer render hardcoded fallback items: an inline `const items: MenuItem[] = [...]` in `Nav` (`components/nav/index.tsx`) and an empty `const items: MenuItem[] = []` in `Footer` (`components/footer/index.tsx`). The components themselves (`QuickLinks`, `MobileMenu`, `Footer`'s `FooterMenu`) already consume `MenuItem[]` in Shopify shape and render up to three levels of nesting. This skill swaps that inline default to a live Shopify menu fetched by handle, keeping the inline items as the fallback if the menu is missing or empty.
 
 ## Before you start
 
@@ -46,14 +46,17 @@ Edit `components/nav/index.tsx`. Add the `getMenu` import and swap the data sour
 import { getMenu } from "@/lib/shopify/operations/menu";
 ```
 
-Inside `Nav`, replace `const items = shopConfig.navigation.nav;` with:
+Inside `Nav`, keep the existing inline default and swap the data source so the Shopify menu takes precedence:
 
 ```tsx
+const defaultItems: MenuItem[] = [
+  { id: "default-nav-shop", title: "Shop", url: "/collections/all", type: "HTTP", items: [] },
+];
 const menu = await getMenu({ handle: "NAV_HANDLE" });
-const items = menu?.items ?? shopConfig.navigation.nav;
+const items = menu?.items ?? defaultItems;
 ```
 
-Replace `"NAV_HANDLE"` with the handle the user provided. `QuickLinks` and `MobileMenu` already accept `MenuItem[]`, so no other component changes are needed. Keep `shopConfig.navigation.nav` as the fallback.
+Replace `"NAV_HANDLE"` with the handle the user provided. `QuickLinks` and `MobileMenu` already accept `MenuItem[]`, so no other component changes are needed. Keep the inline default as the fallback.
 
 ---
 
@@ -67,14 +70,14 @@ Edit `components/footer/index.tsx`. Make the component async, add the `getMenu` 
 import { getMenu } from "@/lib/shopify/operations/menu";
 ```
 
-Change the signature to `async` and replace `const items = shopConfig.navigation.footer;` with:
+Change the signature to `async` and replace `const items: MenuItem[] = [];` with:
 
 ```tsx
 const menu = await getMenu({ handle: "FOOTER_HANDLE" });
-const items = menu?.items ?? shopConfig.navigation.footer;
+const items = menu?.items ?? [];
 ```
 
-Replace `"FOOTER_HANDLE"` with the handle the user provided. `FooterMenu` already accepts `MenuItem[]`. Keep `shopConfig.navigation.footer` as the fallback.
+Replace `"FOOTER_HANDLE"` with the handle the user provided. `FooterMenu` already accepts `MenuItem[]`. The footer's default is empty (no columns), so falling back to `[]` simply hides the section when the Shopify menu is missing or empty.
 
 The `Footer` callsite in `app/layout.tsx` (or wherever it is rendered) already passes `locale`; if it is not wrapped in `<Suspense>`, the menu fetch will block layout render — that is acceptable since `getMenu` is cached with `"use cache: remote"` and `cacheLife("max")`. Wrap in `<Suspense>` only if you specifically want the rest of the page to stream ahead of the footer.
 
@@ -83,7 +86,7 @@ The `Footer` callsite in `app/layout.tsx` (or wherever it is rendered) already p
 ## Guardrails
 
 - The `getMenu()` operation in `lib/shopify/operations/menu.ts` already handles caching (`"use cache: remote"`, `cacheTag("menus")`) and URL transformation. Do not duplicate that logic.
-- Always preserve the `shopConfig.navigation.nav` / `shopConfig.navigation.footer` fallback so a missing or empty Shopify menu doesn't leave the user with a blank nav or footer.
+- Always preserve the inline fallback so a missing or empty Shopify menu doesn't leave the user with a blank nav or footer.
 - External links (URLs starting with `http`) are handled by the existing `MenuLink` helpers in each component — no change needed.
 
 {/* END SKILL CONTENT: enable-shopify-menus */}
