+++ apps/template/.env.example
@@ -1,48 +1,13 @@
-# Required — Shopify Storefront API
+# Required
 SHOPIFY_STORE_DOMAIN="your-store.myshopify.com"
-SHOPIFY_STOREFRONT_ACCESS_TOKEN="your-storefront-access-token"
+SHOPIFY_STOREFRONT_ACCESS_TOKEN="[redacted]"
 
-# Store display name (shown in header, metadata, etc.)
-NEXT_PUBLIC_SITE_NAME="Your Store Name"
-
-# Optional — AI shopping assistant (AI SDK). Overrides shop.config.ts; use "1" or "0".
-# The agent's model routes through the Vercel AI Gateway: on Vercel the project's
-# OIDC token is used automatically; elsewhere set AI_GATEWAY_API_KEY, or run
-# `vercel link` to mint a VERCEL_OIDC_TOKEN for local dev.
-# NEXT_PUBLIC_ENABLE_AGENT="1"
-# AI_GATEWAY_API_KEY="your-ai-gateway-api-key"
-
-# Optional — Vercel BotID protection for POST /api/chat. Defaults to enabled whenever
-# the agent is enabled; set to "0" when self-hosting off Vercel, where checks can't run.
-# Deep Analysis is billed per check on Pro plans; "basic" is free on all plans.
-# NEXT_PUBLIC_ENABLE_BOTID="0"
-# NEXT_PUBLIC_BOTID_CHECK_LEVEL="deepAnalysis"
-
-# Optional — Customer Authentication (Hydrogen + Shopify Customer Account API).
-# The flag overrides shop.config.ts; use "1" or "0".
-# Generate the app-owned cookie encryption secret with `openssl rand -base64 32`.
-# NEXT_PUBLIC_ENABLE_AUTH="1"
-# CUSTOMER_ACCOUNT_SESSION_SECRET="your-random-session-secret"
-# SHOPIFY_CUSTOMER_ACCOUNT_API_CLIENT_ID="your-customer-account-client-id"
-
-# Optional — Base URL for absolute links (sitemap, OG tags, auth callbacks).
-# Defaults to https://${VERCEL_PROJECT_PRODUCTION_URL} when on Vercel.
-# NEXT_PUBLIC_BASE_URL="https://your-domain.com"
-
-# Optional — Shopify webhook signature verification.
-# Required if you wire up /api/webhooks/shopify.
-# SHOPIFY_WEBHOOK_SECRET="your-webhook-secret"
-
-# Optional — CMS draft mode (preview unpublished metaobject content).
-# Required if you use /api/draft for CMS preview links.
-# CMS_DRAFT_MODE_SECRET="your-draft-mode-secret"
-
-# Optional — Override the Shopify Storefront API version. Defaults to unstable.
+# Optional
+# AI_GATEWAY_API_KEY="[redacted]"
+# CUSTOMER_ACCOUNT_SESSION_SECRET="[redacted]"
+# DEBUG_SHOPIFY="1"
+# NEXT_PUBLIC_BASE_URL="your-domain.com"
 # SHOPIFY_API_VERSION="unstable"
-
-# Optional — Verbose Shopify Storefront API request logging in the server console.
-# DEBUG_SHOPIFY="true"
-
-# Optional — UCP agent profile URL for the Storefront MCP catalog search (agent/tools/search_catalog.ts).
-# Only needed if your store requires it; the standard /api/mcp endpoint usually works without it.
+# SHOPIFY_CUSTOMER_ACCOUNT_API_CLIENT_ID="shp_xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx"
+# SHOPIFY_WEBHOOK_SECRET="[redacted]"
 # UCP_AGENT_PROFILE_URL="https://your-agent.example.com/.well-known/ucp"
