+++ apps/template/app/page.tsx
@@ -7,7 +7,7 @@ import { Page } from "@/components/ui/page";
 import { Sections } from "@/components/ui/sections";
 import { getLocale } from "@/lib/params";
 import { buildAlternates, buildOpenGraph } from "@/lib/seo";
-import { shopConfig } from "@/shop.config";
+import { shopConfig } from "@/lib/config";
 
 export async function generateMetadata(): Promise<Metadata> {
   const t = await getTranslations("seo");
@@ -34,7 +34,7 @@ export default async function HomePage() {
     <Page className="pt-0">
       <Sections>
         <section className="grid">
-          <div className="col-start-1 row-start-1 hidden md:block md:aspect-[4/1]" />
+          <div className="col-start-1 row-start-1 hidden md:block md:aspect-4/1" />
           <div className="relative col-start-1 row-start-1 flex items-center justify-center px-5 py-10 lg:px-10">
             <div className="flex flex-col items-center text-center gap-2.5">
               <h1 className="text-3xl md:text-5xl max-w-3xl text-foreground">{t("headline")}</h1>
