+++ apps/template/app/sitemap.xml/route.ts
@@ -1,5 +1,5 @@
 import { getShopifySitemapPagesCount } from "@/lib/shopify/operations/sitemap";
-import { shopConfig } from "@/shop.config";
+import { shopConfig } from "@/lib/config";
 
 function escapeXml(value: string): string {
   return value.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;");
