+++ apps/template/app/sitemap/[shard]/route.ts
@@ -2,7 +2,7 @@ import { notFound } from "next/navigation";
 
 import { getShopPolicies } from "@/lib/shopify/operations/policies";
 import { getShopifySitemapPage, type ShopifySitemapType } from "@/lib/shopify/operations/sitemap";
-import { shopConfig } from "@/shop.config";
+import { shopConfig } from "@/lib/config";
 
 function escapeXml(value: string): string {
   return value.replace(/&/g, "&amp;").replace(/</g, "&lt;").replace(/>/g, "&gt;");
