+++ apps/template/components/schema/site-schema.tsx
@@ -1,4 +1,4 @@
-import { shopConfig } from "@/shop.config";
+import { shopConfig } from "@/lib/config";
 
 function buildSiteSchema(locale: string) {
   return [
