+++ apps/template/lib/markdown/llms.ts
@@ -1,5 +1,5 @@
 import type { Collection } from "@/lib/types";
-import { shopConfig } from "@/shop.config";
+import { shopConfig } from "@/lib/config";
 
 import { escapeMarkdown } from "./utils";
