+++ apps/template/lib/shopify/customer-account.ts
@@ -8,11 +8,11 @@ import {
 } from "@shopify/hydrogen/customer-account";
 
 import { defaultLocale, getCountryCode, getLanguageCode } from "@/lib/i18n";
-import { shopConfig } from "@/shop.config";
+import { shopConfig } from "@/lib/config";
 
 import { resolveShopId } from "./discovery";
 
-const DEBUG = process.env.DEBUG_SHOPIFY === "true";
+const DEBUG = process.env.DEBUG_SHOPIFY === "1";
 
 // Hydrogen requires an HTTPS Origin matching the OAuth-registered auth base URL.
 export async function customerAccountFetch<T>({
