+++ apps/template/lib/shopify/storefront.ts
@@ -12,7 +12,7 @@ import { defaultLocale, getCountryCode, getLanguageCode } from "@/lib/i18n";
 const SHOPIFY_STORE_DOMAIN = process.env.SHOPIFY_STORE_DOMAIN as string;
 const SHOPIFY_ACCESS_TOKEN = process.env.SHOPIFY_STOREFRONT_ACCESS_TOKEN as string;
 const SHOPIFY_API_VERSION = process.env.SHOPIFY_API_VERSION ?? "unstable";
-const DEBUG = process.env.DEBUG_SHOPIFY === "true";
+const DEBUG = process.env.DEBUG_SHOPIFY === "1";
 
 function operationName(body: RequestInit["body"]): string {
   if (typeof body !== "string") return "anonymous";
