+++ turbo.json
@@ -1,17 +1,9 @@
 {
   "$schema": "https://turborepo.dev/schema.json",
   "globalEnv": [
-    "AI_AGENT_DISABLED",
-    "CMS_DRAFT_MODE_SECRET",
     "CUSTOMER_ACCOUNT_SESSION_SECRET",
     "DEBUG_SHOPIFY",
     "NEXT_PUBLIC_BASE_URL",
-    "NEXT_PUBLIC_BOTID_CHECK_LEVEL",
-    "NEXT_PUBLIC_ENABLE_AGENT",
-    "NEXT_PUBLIC_ENABLE_AUTH",
-    "NEXT_PUBLIC_ENABLE_BOTID",
-    "NEXT_PUBLIC_SITE_NAME",
-    "NEXT_PUBLIC_VERCEL_PROJECT_PRODUCTION_URL",
     "SHOPIFY_API_VERSION",
     "SHOPIFY_CUSTOMER_ACCOUNT_API_CLIENT_ID",
     "SHOPIFY_STOREFRONT_ACCESS_TOKEN",
